// crates/ui/web/src/api/types/RepositoryRefBody.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

/**
 * A reference to a `Repository` or `ClusterRepository`.
 */
export type RepositoryRefBody = {
/**
 * `Repository` or `ClusterRepository`.
 */
kind: string,
/**
 * `metadata.name`.
 */
name: string,
/**
 * `metadata.namespace`; omitted for `ClusterRepository`.
 */
namespace?: string | null, };